requirements.txt
Package           Version
----------------- ----------
black             21.11b1
click             8.0.3
colorama          0.4.4
Flask             2.0.2
itsdangerous      2.0.1
Jinja2            3.0.3
MarkupSafe        2.0.1
mypy-extensions   0.4.3
pathspec          0.9.0
pip               21.3.1
platformdirs      2.4.0
python-dotenv     0.19.2
regex             2021.11.10
setuptools        57.4.0
tomli             1.2.2
typing_extensions 4.0.1
Werkzeug          2.0.2
